+++ data/models/ball_model/ball.obj
