+++ scripts/bouncing_ball/config_demo.py
@@ -23,7 +23,7 @@ from ipywidgets import Dropdown, Layout
 from IPython.display import display, HTML, SVG
 
 from pydrake.all import (
-    AddMultibodyPlantSceneGraph, Meshcat, MeshcatVisualizerCpp, DiagramBuilder, 
+    AddMultibodyPlantSceneGraph, Meshcat, MeshcatVisualizer, DiagramBuilder, 
     FindResourceOrThrow, GenerateHtml, InverseDynamicsController, 
     MultibodyPlant, Parser, Simulator, RigidTransform , RotationMatrix )
 from pydrake.multibody.jupyter_widgets import MakeJointSlidersThatPublishOnCallback
@@ -31,63 +31,11 @@ from pydrake.multibody.jupyter_widgets import MakeJointSlidersThatPublishOnCallb
 from pydrake.geometry import (Cylinder, GeometryInstance,
                                 MakePhongIllustrationProperties)
 
+from kinova_arm.utils import AddMultibodyTriad
+
 ##########################
 ## Function Definitions ##
 ##########################
-
-def AddTriad(source_id,
-             frame_id,
-             scene_graph,
-             length=.25,
-             radius=0.01,
-             opacity=1.,
-             X_FT=RigidTransform(),
-             name="frame"):
-    """
-    Adds illustration geometry representing the coordinate frame, with the
-    x-axis drawn in red, the y-axis in green and the z-axis in blue. The axes
-    point in +x, +y and +z directions, respectively.
-    Args:
-      source_id: The source registered with SceneGraph.
-      frame_id: A geometry::frame_id registered with scene_graph.
-      scene_graph: The SceneGraph with which we will register the geometry.
-      length: the length of each axis in meters.
-      radius: the radius of each axis in meters.
-      opacity: the opacity of the coordinate axes, between 0 and 1.
-      X_FT: a RigidTransform from the triad frame T to the frame_id frame F
-      name: the added geometry will have names name + " x-axis", etc.
-    """
-    # x-axis
-    X_TG = RigidTransform(RotationMatrix.MakeYRotation(np.pi / 2),
-                          [length / 2., 0, 0])
-    geom = GeometryInstance(X_FT.multiply(X_TG), Cylinder(radius, length),
-                            name + " x-axis")
-    geom.set_illustration_properties(
-        MakePhongIllustrationProperties([1, 0, 0, opacity]))
-    scene_graph.RegisterGeometry(source_id, frame_id, geom)
-
-    # y-axis
-    X_TG = RigidTransform(RotationMatrix.MakeXRotation(np.pi / 2),
-                          [0, length / 2., 0])
-    geom = GeometryInstance(X_FT.multiply(X_TG), Cylinder(radius, length),
-                            name + " y-axis")
-    geom.set_illustration_properties(
-        MakePhongIllustrationProperties([0, 1, 0, opacity]))
-    scene_graph.RegisterGeometry(source_id, frame_id, geom)
-
-    # z-axis
-    X_TG = RigidTransform([0, 0, length / 2.])
-    geom = GeometryInstance(X_FT.multiply(X_TG), Cylinder(radius, length),
-                            name + " z-axis")
-    geom.set_illustration_properties(
-        MakePhongIllustrationProperties([0, 0, 1, opacity]))
-    scene_graph.RegisterGeometry(source_id, frame_id, geom)
-
-def AddMultibodyTriad(frame, scene_graph, length=.25, radius=0.01, opacity=1.):
-    plant = frame.GetParentPlant()
-    AddTriad(plant.get_source_id(),
-             plant.GetBodyFrameIdOrThrow(frame.body().index()), scene_graph,
-             length, radius, opacity, frame.GetFixedPoseInBodyFrame())
 
 # Building Diagram
 
@@ -97,13 +45,12 @@ plant, scene_graph = AddMultibodyPlantSceneGraph(builder, time_step=1e-4)
 
 # Parse CAD models and add them to the plant.
 # Kinova Gen3 6DoF
-Parser(plant, scene_graph).AddModelFromFile(FindResourceOrThrow("drake/kinova_drake/models/gen3_6dof/urdf/GEN3-6DOF.urdf"))
-# Desk
-Parser(plant, scene_graph).AddModelFromFile("/root/kinova-arm/drake/simpleDesk2/simpleDesk2.urdf")
-# Bouncy Ball
-Parser(plant, scene_graph).AddModelFromFile("/root/kinova-arm/drake/bouncing_ball/ball_model/ball.urdf")
-# Paddle
-Parser(plant, scene_graph).AddModelFromFile("/root/kinova-arm/drake/bouncing_ball/paddle_model/paddle.urdf")
+Parser(plant, scene_graph).AddModelFromFile(
+    "../../data/models/gen3_6dof/urdf/GEN3-6DOF.urdf",
+)
+Parser(plant, scene_graph).AddModelFromFile("../../data/models/simpleDesk/simpleDesk.urdf") # Desk
+Parser(plant, scene_graph).AddModelFromFile("../../data/models/ball_model/ball.urdf") # Bouncy Ball
+Parser(plant, scene_graph).AddModelFromFile("../../data/models/paddle_model/paddle.urdf") # Paddle
 
 """# Ball location
 pusher_position = [0.8,0.5,0.25]
@@ -122,12 +69,12 @@ X_ball.set_translation(ball_position)
 # Weld table to world frame, with rotation about x
 p_RightTableO = [0, 0, 0]
 R_RightTableO = RotationMatrix.MakeXRotation(np.pi/2.0)
-X_WorldTable = RigidTransform(R_RightTableO,p_RightTableO)
+X_WorldTable = RigidTransform(R_RightTableO, p_RightTableO)
 plant.WeldFrames(
     plant.world_frame(), plant.GetFrameByName("simpleDesk"),X_WorldTable)
 
 # Weld robot to table, with translation in x, y and z
-p_PlaceOnTableO = [0.15,0.75,-0.20]
+p_PlaceOnTableO = [0.15, 0.75, -0.20]
 R_PlaceOnTableO = RotationMatrix.MakeXRotation(-np.pi/2.0)
 X_TableRobot = RigidTransform(R_PlaceOnTableO,p_PlaceOnTableO)
 plant.WeldFrames(
@@ -152,14 +99,14 @@ if True:
 
 # Connect to Meshcat
 meshcat0 = Meshcat(port=7001) # Object provides an interface to Meshcat
-mCpp = MeshcatVisualizerCpp(meshcat0)
+mCpp = MeshcatVisualizer(meshcat0)
 mCpp.AddToBuilder(builder,scene_graph,meshcat0)
 
 diagram = builder.Build()
 
 context = diagram.CreateDefaultContext()
 # meshcat.load()
-diagram.Publish(context)
+diagram.ForcedPublish(context)
 
 # Keep connection to meshcat server
 while True:
